IPCs, FBPs, Skrell, Unathi, Vulps, Vox. (#3)
* Adds IPCs and FBPs

FBPs and IPCs are added to Nebula. Also all of the prosthetic sprites are now usable in game.

* Revert "Adds IPCs and FBPs"

This reverts commit 4c74ecc0610f5cab651d6b192d3b64d355de9ea3.

* Adds IPCs and FBPs""

Adds IPCs, FBPs, and the full selection of prosthetics to Nebula. IPCs are listed as "G2 Frames" and will be mechanically slightly different from "G1 Frames" (Utility Frames) and "Shells" (Human looking IPCs)

* Revert "Adds IPCs and FBPs"""

This reverts commit e361832f7ccedfe9f00a0bac205119dbdd896f42.

* Adds IPCs and FBPs""""

Adds IPCs and FBPs as well as the full selection of prosthetics to Nebula.

* Adds IPCs and FBPs

FBPs and IPCs are added to Nebula. Also all of the prosthetic sprites are now usable in game.

* Revert "Adds IPCs and FBPs"

This reverts commit 4c74ecc0610f5cab651d6b192d3b64d355de9ea3.

* Adds IPCs and FBPs""

Adds IPCs, FBPs, and the full selection of prosthetics to Nebula. IPCs are listed as "G2 Frames" and will be mechanically slightly different from "G1 Frames" (Utility Frames) and "Shells" (Human looking IPCs)

* Revert "Adds IPCs and FBPs"""

This reverts commit e361832f7ccedfe9f00a0bac205119dbdd896f42.

* Adds IPCs and FBPs""""

Adds IPCs and FBPs as well as the full selection of prosthetics to Nebula.

* Fixes the absolute mess I made of configs

Puts the config files back in the example folder

* The Race Update

Thanks to Maz for their work on EP5 and for this modpack!

This modpack adds Unathi, Skrell, Vulps, working Vox, and sounds for autohiss. All races are ported directly from baycode, so they should behave (mostly) indentically to their current versions.

* local dumbass forgets to update the #includes

Should all work now

* removal of unrelated content

* Update GeneralContent.dme

* small change to make travis shut up

* makes travis shut up again

* fixes broken music

* TRAVIS PLEASE DONT HURT ME

* Update check-paths.sh

* Fixes Vox stack tech levels so maps don't break